chore: renamed user setting in slice
