+++ frontend/src/components/form/LoginForm.tsx
@@ -15,7 +15,7 @@ import { useNavigate } from "react-router-dom";
 import { yupResolver } from "@hookform/resolvers/yup";
 import AuthAPI from "../../api/auth";
 import Link from "../ui/Link";
-import { login } from "../../feature/user/userSlice";
+import { setUser } from "../../feature/user/userSlice";
 import { UserCredentials, LoginRequest } from "../../proto/user-service";
 import useFixedToast from "../../utils/hooks/useFixedToast";
 import { LOGIN_VALIDATOR } from "../../constants/validators";
@@ -46,7 +46,7 @@ function LoginForm() {
           throw new Error("Something went wrong.");
         }
 
-        dispatch(login({ user }));
+        dispatch(setUser({ user }));
         navigate("/", { replace: true });
       })
       .catch((err) => {
