+++ frontend/src/feature/user/userSlice.ts
@@ -5,38 +5,29 @@ import { RootState } from "../../app/store";
 import { User } from "../../proto/types";
 
 interface UserState {
-  // sessionToken: string;
-  user?: User;
+	user?: User;
 }
 
 const initialState: UserState = {
-  // sessionToken: "",
-  user: undefined,
+	user: undefined,
 };
 
 export const userSlice = createSlice({
-  name: "user",
-  initialState,
-  reducers: {
-    login: (state, action: PayloadAction<Required<UserState>>) => {
-      const { user } = action.payload;
-      // state.sessionToken = sessionToken;
-      state.user = user;
-    },
-    logout: (state) => {
-      // state.sessionToken = initialState.sessionToken;
-      state.user = initialState.user;
-    },
-    changeNickname: (state, action: PayloadAction<Required<UserState>>) => {
-      const { user } = action.payload;
-      // state.sessionToken = sessionToken;
-      state.user = user;
-    },
-  },
+	name: "user",
+	initialState,
+	reducers: {
+		setUser: (state, action: PayloadAction<Required<UserState>>) => {
+			const { user } = action.payload;
+			state.user = user;
+		},
+		logout: (state) => {
+			state.user = initialState.user;
+		},
+	},
 });
 
 export const selectUser = (state: RootState) => state.user.user;
 
-export const { login, logout, changeNickname } = userSlice.actions;
+export const { setUser, logout } = userSlice.actions;
 
 export default userSlice.reducer;
