+++ frontend/src/pages/Session.tsx
@@ -17,7 +17,7 @@ import SessionNavbar from "../components/ui/navbar/SessionNavbar";
 import Editor from "../components/editor/Editor";
 import useFixedToast from "../utils/hooks/useFixedToast";
 import { selectUser } from "../feature/user/userSlice";
-import { Chat, Language } from "../types";
+import { Chat, Language } from "../types/types";
 import { Question } from "../proto/types";
 import saveFile from "../utils/fileDownloadUtil";
 import {
@@ -28,7 +28,7 @@ import {
   changeWSStatus,
   reset,
   setExecution,
-  setQuestion,
+  setQuestion
 } from "../feature/session/sessionSlice";
 import FontSizeSelector from "../components/editor/FontSizeSelector";
 
@@ -70,7 +70,7 @@ function Session() {
         yd,
         {
           params,
-          disableBc: true,
+          disableBc: true
         }
       );
 
@@ -101,13 +101,13 @@ function Session() {
 
       ws.on("user_join", (joinedNickname: Nickname) => {
         toast.sendSuccessMessage("", {
-          title: `${joinedNickname.nickname} has joined the room!`,
+          title: `${joinedNickname.nickname} has joined the room!`
         });
       });
 
       ws.on("user_leave", (leftNickname: Nickname) => {
         toast.sendAlertMessage("", {
-          title: `${leftNickname.nickname} has left the room.`,
+          title: `${leftNickname.nickname} has left the room.`
         });
       });
 
@@ -159,7 +159,7 @@ function Session() {
       // Yjs initialisation
       const tempyDoc = new Y.Doc();
       const params: { [x: string]: string } = {
-        room: roomToken === undefined ? "" : roomToken,
+        room: roomToken === undefined ? "" : roomToken
       };
 
       const tempprovider = buildWSProvider(tempyDoc, params);
